Fix breaking Select docs
