+++ packages/grafana-ui/src/components/Forms/Select/Select.story.tsx
@@ -9,41 +9,12 @@ import { Button } from '../Button';
 import { ButtonSelect } from './ButtonSelect';
 import { getIconKnob } from '../../../utils/storybook/knobs';
 import kebabCase from 'lodash/kebabCase';
+import { generateOptions } from './mockOptions';
 
 export default {
   title: 'Forms/Select',
   component: Select,
   decorators: [withCenteredStory, withHorizontallyCenteredStory],
-};
-
-export const generateOptions = () => {
-  const values = [
-    'Sharilyn Markowitz',
-    'Naomi Striplin',
-    'Beau Bevel',
-    'Garrett Starkes',
-    'Hildegarde Pedro',
-    'Gudrun Seyler',
-    'Eboni Raines',
-    'Hye Felix',
-    'Chau Brito',
-    'Heidy Zook',
-    'Karima Husain',
-    'Virgil Mckinny',
-    'Kaley Dodrill',
-    'Sharan Ruf',
-    'Edgar Loveland',
-    'Judie Sanger',
-    'Season Bundrick',
-    'Ok Vicente',
-    'Garry Spitz',
-    'Han Harnish',
-  ];
-
-  return values.map<SelectableValue<string>>(name => ({
-    value: kebabCase(name),
-    label: name,
-  }));
 };
 
 const loadAsyncOptions = () => {
